(MOT-4253) refactor(harness): split reactive E2E scenario
